+++ src/patch/core/SharedDef.zig
@@ -27,7 +27,7 @@ const Trig = r.Entity.Trig.Trig;
 const ModelTriggerDescription = r.Model.ModelTriggerDescription;
 const TextDef = r.Text.TextDef;
 
-const RaceState = enum(u8) { None, PreRace, Countdown, Racing, PostRace, PostRaceExiting };
+pub const RaceState = enum(u8) { None, PreRace, Countdown, Racing, PostRace, PostRaceExiting };
 
 pub const GLOBAL_STATE_VERSION = 6;
 
@@ -77,7 +77,7 @@ pub const GlobalState = extern struct {
     } = .{},
 };
 
-pub const GLOBAL_FUNCTION_VERSION = 29;
+pub const GLOBAL_FUNCTION_VERSION = 30;
 
 // TODO: fnptr for nullable handles, or handles in general?
 pub const GlobalFunction = extern struct {
@@ -147,4 +147,32 @@ pub const GlobalFunction = extern struct {
     ) callconv(.C) Handle(u16),
     RTriggerRelease: *const fn (Handle(u16)) callconv(.C) void,
     RTriggerReleaseAll: *const fn () callconv(.C) void,
+    // State;  previously accessed via 'global state'
+    SPatchMemory: *const fn () callconv(.C) [*]u8, // patch_memory
+    SPatchSize: *const fn () callconv(.C) u32, // patch_size
+    SPatchOffset: *const fn () callconv(.C) u32, // patch_offset, WARN: some funcs write to this
+    SInitLatePassed: *const fn () callconv(.C) bool, // init_late_passed
+    SPracticeMode: *const fn () callconv(.C) bool, // practice_mode
+    SWindowInForeground: *const fn () callconv(.C) bool, // window_in_foreground
+    SDt: *const fn () callconv(.C) f32, // dt_f
+    SFPS: *const fn () callconv(.C) f32, // fps
+    SFPSAvg: *const fn () callconv(.C) f32, // fps_avg
+    STimestamp: *const fn () callconv(.C) u32, // timestamp
+    SFrameCount: *const fn () callconv(.C) u32, // frame_count
+    SInRace: *const fn () callconv(.C) ActiveState, // in_race
+    SRaceState: *const fn () callconv(.C) RaceState, // race_state
+    SRaceStatePrev: *const fn () callconv(.C) RaceState, // race_state_prev
+    SRaceStateNew: *const fn () callconv(.C) bool, // race_state_new
+    SPlayerUpgrades: *const fn () callconv(.C) bool, // player -> upgrades
+    SPlayerUpgradesLv: *const fn (out: [*]u8) callconv(.C) void, // 7-byte array; player -> upgrades_lv
+    SPlayerUpgradesHP: *const fn (out: [*]u8) callconv(.C) void, // 7-byte array; player -> upgrades_hp
+    SPlayerFlags1: *const fn () callconv(.C) TestFlags1, // player -> flags1
+    SPlayerBoosting: *const fn () callconv(.C) ActiveState, // player -> boosting
+    SPlayerUnderheating: *const fn () callconv(.C) ActiveState, // player -> underheating
+    SPlayerOverheating: *const fn () callconv(.C) ActiveState, // player -> overheating
+    SPlayerDead: *const fn () callconv(.C) ActiveState, // player -> dead
+    SPlayerDeaths: *const fn () callconv(.C) u32, // player -> deaths
+    SPlayerHeatRate: *const fn () callconv(.C) f32, // player -> heat_rate
+    SPlayerCoolRate: *const fn () callconv(.C) f32, // player -> cool_rate
+    SPlayerHeat: *const fn () callconv(.C) f32, // player -> heat
 };